+++ CHANGES.md
@@ -16,6 +16,7 @@ Changes in 4.7
 - Add CLASS_RANK_CALCULATE_MPS to CONFIG table in Update.fnc.php & rosariosis.sql
 - Add ClassRankMaybeCalculate(), ClassRankCalculateAJAX() & ClassRankCalculateAddMP() in ClassRank.inc.php
 - Automatic Class Rank calculation in EditReportCardGrades.php, GPARankList.php, InputFinalGrades.php, Transcripts.php
+- SQL performance: rewrite set_class_rank_mp() function in rosariosis.sql & Update.fnc.php
 
 Changes in 4.6.2
 ----------------